+++ src/main/java/ru/kelcuprum/pplhelper/gui/TextureHelper.java
@@ -25,13 +25,13 @@ import java.nio.file.attribute.BasicFileAttributes;
 import java.util.HashMap;
 
 import static ru.kelcuprum.pplhelper.PepelandHelper.Icons.PACK_INFO;
+import static ru.kelcuprum.pplhelper.PepelandHelper.Icons.WHITE_PEPE;
 
 public class TextureHelper {
     public static HashMap<String, ResourceLocation> resourceLocationMap = new HashMap<>();
     public static HashMap<String, Boolean> urls = new HashMap<>();
     public static HashMap<String, DynamicTexture> urlsTextures = new HashMap<>();
     public static HashMap<String, NativeImage> urlsImages = new HashMap<>();
-    public static JsonArray map = new JsonArray();
     // Internet
     public static ResourceLocation getTexture(String url, String id) {
         id = formatUrls(id.toLowerCase());
@@ -42,7 +42,7 @@ public class TextureHelper {
                 String finalId = id;
                 new Thread(() -> registerTexture(url, finalId, AlinLib.MINECRAFT.getTextureManager(), GuiUtils.getResourceLocation("pplhelper", finalId))).start();
             }
-            return PACK_INFO;
+            return WHITE_PEPE;
         }
     }
 
@@ -51,27 +51,20 @@ public class TextureHelper {
         PepelandHelper.LOG.debug(String.format("REGISTER: %s %s", url, id));
         DynamicTexture texture;
         if (urlsTextures.containsKey(url)) {
-            addToMap(id, url);
             texture = urlsTextures.get(url);
         } else {
             NativeImage image;
-            File textureFile = getTextureFile(id);
-            boolean isFileExists = textureFile.exists();
             try {
-                BufferedImage bufferedImage = isFileExists ? ImageIO.read(getTextureFile(id)) : ImageIO.read(new URL(url));
+                BufferedImage bufferedImage = ImageIO.read(new URL(url));
                 ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
                 ImageIO.write(bufferedImage, url.contains(".webp") ? "webp" : "png", byteArrayOutputStream);
                 InputStream is = new ByteArrayInputStream(byteArrayOutputStream.toByteArray());
                 image = NativeImage.read(is);
-                if (!isFileExists) {
-                    Files.createDirectories(textureFile.toPath().getParent());
-                    Files.write(textureFile.toPath(), byteArrayOutputStream.toByteArray());
-                }
                 texture = new DynamicTexture(image);
                 urlsTextures.put(url, texture);
                 urlsImages.put(url, image);
             } catch (Exception e) {
-                PepelandHelper.LOG.log("Error loading image from URL: " + url + " - " + e.getMessage());
+                PepelandHelper.LOG.error("Error loading image from URL: " + url + " - " + e.getMessage());
                 resourceLocationMap.put(id, PACK_INFO);
                 return;
             }
@@ -79,7 +72,6 @@ public class TextureHelper {
         if (textureManager != null) {
             textureManager.register(textureId, texture);
             resourceLocationMap.put(id, textureId);
-            addToMap(id, url);
         }
     }
 
@@ -92,28 +84,25 @@ public class TextureHelper {
                 String finalId = id;
                 new Thread(() -> registerBanner(url, finalId, AlinLib.MINECRAFT.getTextureManager(), GuiUtils.getResourceLocation("pplhelper", finalId))).start();
             }
-            return PACK_INFO;
+            return WHITE_PEPE;
         }
     }
 
     @Async.Execute
     public static void registerBanner(String url, String id, TextureManager textureManager, ResourceLocation textureId) {
-        PepelandHelper.LOG.log(String.format("REGISTER: %s %s", url, id), Level.DEBUG);
+        PepelandHelper.LOG.debug(String.format("REGISTER: %s %s", url, id));
         DynamicTexture texture;
-        if (urlsTextures.containsKey(url)) {
-            addToMap(id, url);
+        if (urlsTextures.containsKey(url))
             texture = urlsTextures.get(url);
-        } else {
+        else {
             NativeImage image;
-            File textureFile = getTextureFile(id);
-            boolean isFileExists = textureFile.exists();
             try {
-                BufferedImage bufferedImage = isFileExists ? ImageIO.read(getTextureFile(id)) : ImageIO.read(new URL(url));
+                BufferedImage bufferedImage = ImageIO.read(new URL(url));
                 double widthScale = (double) bufferedImage.getWidth() / 750;
                 int threeHundredBucks = (int) (300 * widthScale);
                 double scale = (double) bufferedImage.getHeight() / threeHundredBucks;
                 int height = (int) (bufferedImage.getHeight() / scale);
-                if (bufferedImage.getHeight() > height && !isFileExists) {
+                if (bufferedImage.getHeight() > height) {
                     int y = (bufferedImage.getHeight() - height) / 2;
                     bufferedImage = bufferedImage.getSubimage(0, y, bufferedImage.getWidth(), bufferedImage.getHeight()-(y*2));
                 }
@@ -121,15 +110,11 @@ public class TextureHelper {
                 ImageIO.write(bufferedImage, "png", byteArrayOutputStream);
                 InputStream is = new ByteArrayInputStream(byteArrayOutputStream.toByteArray());
                 image = NativeImage.read(is);
-                if (!isFileExists) {
-                    Files.createDirectories(textureFile.toPath().getParent());
-                    Files.write(textureFile.toPath(), byteArrayOutputStream.toByteArray());
-                }
                 texture = new DynamicTexture(image);
                 urlsTextures.put(url, texture);
                 urlsImages.put(url, image);
             } catch (Exception e) {
-                PepelandHelper.LOG.log("Error loading image from URL: " + url + " - " + e.getMessage());
+                PepelandHelper.LOG.error("Error loading image from URL: " + url + " - " + e.getMessage());
                 resourceLocationMap.put(id, PACK_INFO);
                 return;
             }
@@ -137,67 +122,7 @@ public class TextureHelper {
         if (textureManager != null) {
             textureManager.register(textureId, texture);
             resourceLocationMap.put(id, textureId);
-            addToMap(id, url);
         }
-    }
-
-    public static void addToMap(String id, String url){
-        JsonObject data = new JsonObject();
-        data.addProperty("url", url);
-        data.addProperty("id", id);
-        if (!map.contains(data)) map.add(data);
-    }
-
-    public static File getTextureFile(String url) {
-        return new File("./config/pplhelper/textures/" + url + ".png");
-    }
-
-    public static void saveMap() {
-        try {
-            Path path = new File("./config/pplhelper/textures/map.json").toPath();
-            Files.createDirectories(path.getParent());
-            Files.writeString(path, map.toString());
-        } catch (IOException e) {
-            PepelandHelper.LOG.log(e.getLocalizedMessage(), Level.ERROR);
-        }
-    }
-
-    public static void loadTextures(TextureManager textureManager) {
-        loadMap();
-        try {
-            final JsonArray finalMap = map;
-            for (JsonElement json : finalMap) {
-                JsonObject data = json.getAsJsonObject();
-                ResourceLocation l = GuiUtils.getResourceLocation("pplhelper", data.get("id").getAsString());
-                registerTexture(data.get("url").getAsString(), data.get("id").getAsString(), textureManager, l);
-            }
-        } catch (Exception e) {
-            PepelandHelper.LOG.log("MAP ERROR!", Level.ERROR);
-        }
-    }
-
-    public static void loadMap() {
-        File mapFile = new File("./config/pplhelper/textures/map.json");
-        JsonArray finalMap;
-        if (mapFile.exists() && mapFile.isFile()) {
-            try {
-                finalMap = GsonHelper.parseArray(Files.readString(mapFile.toPath()));
-            } catch (Exception e) {
-                finalMap = new JsonArray();
-                PepelandHelper.LOG.log(e.getMessage() == null ? e.getClass().getName() : e.getMessage(), Level.ERROR);
-            }
-        } else finalMap = new JsonArray();
-        JsonArray array = new JsonArray();
-        for(JsonElement element : finalMap){
-            JsonObject jsonObject = element.getAsJsonObject();
-            File file = getTextureFile(jsonObject.get("id").getAsString());
-            try {
-                if (file.exists() && System.currentTimeMillis() - Files.readAttributes(file.toPath(), BasicFileAttributes.class).creationTime().toMillis() >= 259200000)
-                    file.deleteOnExit();
-                else array.add(element);
-            } catch (Exception ignored){}
-        }
-        map = array;
     }
 
     public static String formatUrls(String url) {